[Decomp][Spell] Cohesion split: extract heal/power/item-create handlers to SpellEffectHealPower.cpp
Move the first contiguous block of effect handlers (EffectApplyAura
through EffectEnergisePct, plus the DoCreateItem helper) out of
SpellEffects.cpp into a new SpellEffectHealPower.cpp translation unit.

Pure file-level cohesion split: declarations stay in Spell.h, the
SpellEffects[] dispatch table stays in SpellEffects.cpp and resolves the
moved member methods by address at link time. Byte-identical move, no
behaviour change. SpellEffects.cpp 6027 -> 5224.
